+++ ostap/plotting/canvas.py
@@ -19,13 +19,16 @@ __all__     = (
     'canvas_pull'      , ## split canvas into two pads with no vertical interspace
     'draw_pads'        , ## plot sequence of object on sequence of pads, adjustinng axis label size
     'AutoPlots'        , ## context manager to activate the auto-plotting machinery
-    'auto_plots'       , ## ditto, but as function 
+    'auto_plots'       , ## ditto, but as function
+    'use_pad'          , ## context manager to modifty TPad 
     )
 # =============================================================================
 import ROOT, os, tempfile  
 import ostap.core.core
 import ostap.plotting.style
 from   sys import version_info as python_version
+from   ostap.utils.cidict import cidict
+from   ostap.core.core    import cidict_fun
 # =============================================================================
 # logging 
 # =============================================================================
@@ -310,7 +313,10 @@ if not hasattr ( ROOT.TObject , 'draw_with_autoplot' ) :
 
         if ROOT.gPad :
             plot = AutoPlots.plot()
-            if plot : ROOT.gPad >> plot
+            if plot :
+                cnv = ROOT.gPad.GetCanvas ()
+                if cnv : 
+                    cnv >> plot
             
         return result
     
@@ -357,7 +363,7 @@ def canvas_partition ( canvas                        ,
                        left_margin   = margin_left   , 
                        right_margin  = margin_right  , 
                        bottom_margin = margin_bottom , 
-                       top_margin    = margin_right  ,
+                       top_margin    = margin_top    ,
                        hSpacing      = 0.0           ,
                        vSpacing      = 0.0           ) :
     """Perform partition of Canvas into pads with no inter-margins
@@ -744,8 +750,149 @@ def draw_pads ( objects            ,
         obj.draw ()
         
         if c : c.cd(0)
+
+# =============================================================================
+## change main parametes of TAttPad
+def set_pad ( pad , **config ) :
+    """Change main parametes of `TAttPad`"""
+
+    conf = cidict ( transform = cidict_fun )
+    conf.update ( config )
+
+    changed = {}
+
+    if 'frame_border_mode' in conf :
+        changed[ 'frame_border_mode' ] = pad.GetFrameBorderMode()
+        pad.SetFrameBorderMode ( conf.pop ('frame_border_mode') )
+
+    if 'frame_border_size' in conf :
+        changed[ 'frame_border_size' ] = pad.GetFrameBorderSize ()
+        pad.SetFrameBorderSize ( conf.pop ('frame_border_size') )
         
+    if 'frame_fill_color' in conf :
+        changed[ 'frame_fill_color' ] = pad.GetFrameFillColor ()
+        pad.SetFrameFillColor ( conf.pop ('frame_fill_color') )
+
+    if 'frame_fill_style' in conf :
+        changed[ 'frame_fill_style' ] = pad.GetFrameFillStyle ()
+        pad.SetFrameFillStyle ( conf.pop ('frame_fill_style') )
+
+    if 'frame_line_color' in conf :
+        changed[ 'frame_line_color' ] = pad.GetFrameLineColor ()
+        pad.SetFrameLineColor ( conf.pop ('frame_line_color') )
+
+    if 'frame_line_style' in conf :
+        changed[ 'frame_line_style' ] = pad.GetFrameLineStyle ()
+        pad.SetFrameLineStyle ( conf.pop ('frame_line_style') )
+
+    if 'frame_line_width' in conf :
+        changed[ 'frame_line_width' ] = pad.GetFrameLineWidth ()
+        pad.SetFrameLineWidth ( conf.pop ('frame_line_width') )
+
+    if 'a_file' in conf :
+        changed[ 'a_file' ] = pad.GetAfile ()
+        pad.SetAfile ( conf.pop ('a_file') )
+
+    if 'a_stat' in conf :
+        changed[ 'a_stat' ] = pad.GetAstat ()
+        pad.SetAstat ( conf.pop ('a_stat') )
+
+    if 'x_file' in conf :
+        changed[ 'x_file' ] = pad.GetXfile ()
+        pad.SetXfile ( conf.pop ('x_file') )
+
+    if 'x_stat' in conf :
+        changed[ 'x_stat' ] = pad.GetXstat ()
+        pad.SetXstat ( conf.pop ('x_stat') )
+
+    if 'y_file' in conf :
+        changed[ 'y_file' ] = pad.GetYfile ()
+        pad.SetYfile ( conf.pop ('y_file') )
+
+    if 'y_stat' in conf :
+        changed[ 'y_stat' ] = pad.GetYstat ()
+        pad.SetYstat ( conf.pop ('y_stat') )
+
+    
+    if 'top_margin' in conf or 'margin_top' in conf :                
+        changed ['margin_top']  = pad.GetTopMargin()
+        if 'top_margin' in conf    : pad.SetTopMargin    ( conf.pop ( 'top_margin'   ) )
+        else                       : pad.SetTopMargin    ( conf.pop ( 'margin_top'   ) )
         
+    if 'bottom_margin' in conf or 'margin_bottom' in conf :            
+        changed ['margin_bottom']  = pad.GetBottomMargin()
+        if 'bottom_margin' in con  : pad.SetBottomMargin ( conf.pop ( 'bottom_margin' ) )
+        else                       : pad.SetBottomMargin ( conf.pop ( 'margin_bottom' ) ) 
+        
+    if 'left_margin' in conf or 'margin_left' in conf :                
+        changed ['margin_left']  = pad.GetLeftMargin()
+        if 'left_margin' in conf   : pad.SetLeftMargin   ( conf.pop ( 'left_margin'   ) )
+        else                       : pad.SetLeftMargin   ( conf.pop ( 'margin_left'   ) )
+        
+    if 'right_margin' in conf or 'margin_right' in conf :                
+        changed ['margin_right']  = pad.GetRightMargin()
+        if 'right_margin' in conf  : pad.SetRightMargin  ( conf.pop ( 'right_margin'  ) )
+        else                       : pad.SetRightMargin  ( conf.pop ( 'margin_right'  ) )
+        
+    if conf :
+        logger.warning ("set_pad: unprocessed items: %s" % conf ) 
+
+    return changed 
+    
+# =============================================================================
+## helper context manager for <code>TAttPad</code> objects
+#  @see TAttPad 
+class UsePad(object) :
+    """Helper context manager for `TAttPad` objects
+    - see `TAttPad`
+    """
+
+    def  __init__ ( self , pad = None , **config ) :
+
+        self.__pad     = pad if ( pad and isinstance ( pad , ROOT.TAttPad ) ) else None 
+        self.__config  = config
+        self.__changed = {} 
+        
+    def __enter__ ( self ) :
+            
+        if not self.__pad and ROOT.gPad : self.__pad = ROOT.gPad
+        
+        if self.pad : 
+            self.__changed = set_pad ( self.pad , **self.config )
+
+        return self 
+        
+    def __exit__ ( self , *_ ) :
+
+        if self.pad and self.changed :
+            set_pad ( self.pad , **self.changed )
+
+    @property
+    def pad ( self ) :
+        """``pad'' : pad to be configured"""
+        return self.__pad
+    
+    @property
+    def config ( self ) :
+        """``config'' : cofiguration pad to be"""
+        return self.__config
+
+    @property
+    def changed ( self ) :
+        """``changed'' : changed papameters"""
+        return self.__changed
+    
+# =============================================================================
+## helpoer context manager to modify <code>TAttPad</code>
+#  @see TAttPad 
+def use_pad ( pad , **config ) :
+    """Helper context manager for `TAttPad` objects
+    - see `TAttPad`
+    """
+    return UsePad ( pad , **config ) 
+
+usePad = use_pad
+
 # =============================================================================
 _decorated_classes_  = (
     ROOT.TVirtualPad , 
@@ -767,5 +914,5 @@ if '__main__' == __name__ :
     docme ( __name__ , logger = logger )
 
 # =============================================================================
-# The END 
+##                                                                      The END 
 # =============================================================================